Herencia de roles implementada Clase padre: Roles Clases hijas: Administrator, User Clase vieja eliminada: Auth
